website_chatbot/requirements.txt
# Website Chatbot Requirements

# Microsoft Agent Framework (core framework)
agent-framework>=0.1.0

# Google Cloud AI Platform
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Web scraping and crawling
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0

# Vector search and embeddings
chromadb>=0.4.22
chromadb-client>=0.4.22
numpy>=1.24.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Async support
asyncio>=3.4.3

# Logging and debugging
coloredlogs>=15.0
